feat: auto revalidationを使用して全員承認を判定した実行ボタンの制御を行う
