+++ src/components/atoms/Approvers/Approvers.stories.ts
@@ -1,5 +1,4 @@
 import { Meta, StoryObj } from '@storybook/react'
-import { within } from '@storybook/testing-library'
 
 import { Approvers } from './Approvers'
 
@@ -17,7 +16,7 @@ export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement)
+  args: {
+    agreements: [],
   },
 }
